refactor(compiler): split entry points from legacy helpers in control-flow emit
Before: `ir-to-client-js/emit-control-flow.ts` carried two roles:
  - public entry points consumed by `generate-init.ts`
  - leaf stringify-layer helpers consumed by `control-flow/{plan,stringify}/*`

The result was a circular import shape: control-flow/* imported back
into the same file that imported control-flow/*. TypeScript handled it
via type-only fields, but it obscured the actual dependency direction.

This PR splits the file along that natural seam:

  ir-to-client-js/
  ├── control-flow.ts                 # NEW (102 lines)
  │   = 3 public entry points + 2 dispatcher wrappers, all thin
  │     calls into the Plan layer.
  └── control-flow/
      ├── legacy-helpers.ts           # MOVED from emit-control-flow.ts
      │   = stringify-layer helpers shared by Plan stringifiers.
      ├── plan/...                    # imports from `../legacy-helpers`
      └── stringify/...               # imports from `../legacy-helpers`

Dependency direction is now one-way:

  control-flow.ts -> control-flow/{plan,stringify}/* -> legacy-helpers.ts

Other adjustments:
- `emitCompositeBranchLoop` (a 1-line wrapper) is inlined into its
  only caller `emitBranchLoopBody` — one fewer indirection.
- Inline `import('./types')` references for `IRProp` / `IRNode` are
  fixed to point at `'../../types'` (jsx/src/types.ts) — the legacy
  paths happened to compile, but the new file location surfaced them.

`emit-control-flow.ts` is renamed to `control-flow/legacy-helpers.ts`
via `git mv` so blame / history follow the larger chunk (helpers).

Verified: jsx 766 / 766, adapter-tests 214 / 214, client 212 / 212;
clean `bun run --filter '@barefootjs/jsx' build` exits 0. Output is
byte-identical to main.
