+++ packages/jsx/src/ir-to-client-js/control-flow/plan/build-component-loop.ts
@@ -27,7 +27,7 @@ import {
   buildComponentPropsExpr,
   buildCompSelector,
   isTextOnlyConditional,
-} from '../../emit-control-flow'
+} from '../legacy-helpers'
 import { irChildrenToJsExpr } from '../../html-template'
 import type { ComponentLoopPlan, NestedComponentInit } from './types'
